requirements.txt
config==0.5.0.post0
discord.py==1.7.1